fixes #13 implement mode instead of input panel
Significant refactoring that puts ST2 into a new mode when the EasyMotion shortcut is hit.  It disables vintage command mode (if active) and waits for new keys from the user.

Some additional testing and a few bugs on the edge (especially allow escape/ctrl-c to cancel the mode) need to be worked through before this can be merged to master.